Clean up controllers by moving all Form logic to dedicated classes
gitignore phpDocumentor files (including pubkey, if present)
