+++ .gitignore
@@ -23,5 +23,4 @@
 *.swp
 .DS_Store
 i18n_helper.js
-phpDocumentor.phar
-
+phpDocumentor.phar*
