+++ app/Resources/views/events/show.html.twig
@@ -1,6 +1,6 @@
 {% extends 'base.html.twig' %}
 {% import 'macros/layout.html.twig' as layout %}
-{% form_theme form 'blocks/forms.html.twig' %}
+{% form_theme participantForm 'blocks/forms.html.twig' %}
 
 {% block breadcrumb %}
     <ol class="breadcrumb">
@@ -118,7 +118,7 @@
         </section><!-- #stats -->
 
         {# Participants accordian #}
-        {% include 'events/_participants.html.twig' with {'form': form, 'event': event, 'isOrganizer': isOrganizer} %}
+        {% include 'events/_participants.html.twig' with {'form': participantForm, 'event': event, 'isOrganizer': isOrganizer} %}
 
     </div>
 {% endblock %}
