+++ public/assets/app.e04b02e9.js
@@ -7,4 +7,4 @@
  *
  * Modifications by Paul Warelis and Alexey Gordeyev
  */
-!function(n){"use strict";var e=function(e,t){n.fn.typeahead.defaults,t.scrollBar&&(t.items=100,t.menu='<ul class="typeahead dropdown-menu" style="max-height:220px;overflow:auto;"></ul>');var a=this;if(a.$element=n(e),a.options=n.extend({},n.fn.typeahead.defaults,t),a.$menu=n(a.options.menu).insertAfter(a.$element),a.eventSupported=a.options.eventSupported||a.eventSupported,a.grepper=a.options.grepper||a.grepper,a.highlighter=a.options.highlighter||a.highlighter,a.lookup=a.options.lookup||a.lookup,a.matcher=a.options.matcher||a.matcher,a.render=a.options.render||a.render,a.onSelect=a.options.onSelect||null,a.sorter=a.options.sorter||a.sorter,a.select=a.options.select||a.select,a.source=a.options.source||a.source,a.displayField=a.options.displayField||a.displayField,a.valueField=a.options.valueField||a.valueField,a.autoSelect=a.options.autoSelect||a.autoSelect,a.options.ajax){var r=a.options.ajax;"string"==typeof r?a.ajax=n.extend({},n.fn.typeahead.defaults.ajax,{url:r}):("string"==typeof r.displayField&&(a.displayField=a.options.displayField=r.displayField),"string"==typeof r.valueField&&(a.valueField=a.options.valueField=r.valueField),a.ajax=n.extend({},n.fn.typeahead.defaults.ajax,r)),a.ajax.url||(a.ajax=null),a.query=""}else a.source=a.options.source,a.ajax=null;a.shown=!1,a.listen()};e.prototype={constructor:e,eventSupported:function(n){var e=n in this.$element;return e||(this.$element.setAttribute(n,"return;"),e="function"==typeof this.$element[n]),e},select:function(){var n=this.$menu.find(".active");if(n.length){var e=n.attr("data-value"),t=this.$menu.find(".active a").text();this.$element.val(this.updater(t)).change(),this.options.onSelect&&this.options.onSelect({value:e,text:t})}return this.hide()},updater:function(n){return n},show:function(){var e=n.extend({},this.$element.position(),{height:this.$element[0].offsetHeight});if(this.$menu.css({top:e.top+e.height,left:e.left}),this.options.alignWidth){var t=n(this.$element[0]).outerWidth();this.$menu.css({width:t})}return this.$menu.show(),this.shown=!0,this},hide:function(){return this.$menu.hide(),this.shown=!1,this},ajaxLookup:function(){function e(){this.ajaxToggleLoadClass(!0),this.ajax.xhr&&this.ajax.xhr.abort();var e=this.ajax.preDispatch?this.ajax.preDispatch(t):{query:t};this.ajax.xhr=n.ajax({url:this.ajax.url,data:e,success:n.proxy(this.ajaxSource,this),type:this.ajax.method||"get",dataType:"json",headers:this.ajax.headers||{}}),this.ajax.timerId=null}var t=n.trim(this.$element.val());return t===this.query?this:(this.query=t,this.ajax.timerId&&(clearTimeout(this.ajax.timerId),this.ajax.timerId=null),!t||t.length<this.ajax.triggerLength?(this.ajax.xhr&&(this.ajax.xhr.abort(),this.ajax.xhr=null,this.ajaxToggleLoadClass(!1)),this.shown?this.hide():this):(this.ajax.timerId=setTimeout(n.proxy(e,this),this.ajax.timeout),this))},ajaxSource:function(n){this.ajaxToggleLoadClass(!1);var e,t=this;if(t.ajax.xhr)return t.ajax.preProcess&&(n=t.ajax.preProcess(n)),t.ajax.data=n,e=t.grepper(t.ajax.data)||[],e.length?(t.ajax.xhr=null,t.render(e.slice(0,t.options.items)).show()):t.shown?t.hide():t},ajaxToggleLoadClass:function(n){this.ajax.loadingClass&&this.$element.toggleClass(this.ajax.loadingClass,n)},lookup:function(n){var e,t=this;return t.ajax?void t.ajaxer():(t.query=t.$element.val(),t.query&&(e=t.grepper(t.source))?(0==e.length&&(e[0]={id:-21,name:"Result not Found"}),t.render(e.slice(0,t.options.items)).show()):t.shown?t.hide():t)},matcher:function(n){return~n.toLowerCase().indexOf(this.query.toLowerCase())},sorter:function(n){if(this.options.ajax)return n;for(var e,t=[],a=[],r=[];e=n.shift();)e.toLowerCase().indexOf(this.query.toLowerCase())?~e.indexOf(this.query)?a.push(e):r.push(e):t.push(e);return t.concat(a,r)},highlighter:function(n){var e=this.query.replace(/[\-\[\]{}()*+?.,\\\^$|#\s]/g,"\\$&");return n.replace(new RegExp("("+e+")","ig"),function(n,e){return"<strong>"+e+"</strong>"})},render:function(e){var a,r=this,i="string"==typeof r.options.displayField;return e=n(e).map(function(e,o){return"object"==(void 0===o?"undefined":t(o))?(a=i?o[r.options.displayField]:r.options.displayField(o),e=n(r.options.item).attr("data-value",o[r.options.valueField])):(a=o,e=n(r.options.item).attr("data-value",o)),e.find("a").html(r.highlighter(a)),e[0]}),r.autoSelect&&e.first().addClass("active"),this.$menu.html(e),this},grepper:function(e){var t,a,r=this,i="string"==typeof r.options.displayField;if(!(i&&e&&e.length))return null;if(e[0].hasOwnProperty(r.options.displayField))t=n.grep(e,function(n){return a=i?n[r.options.displayField]:r.options.displayField(n),r.matcher(a)});else{if("string"!=typeof e[0])return null;t=n.grep(e,function(n){return r.matcher(n)})}return this.sorter(t)},next:function(e){var t=this.$menu.find(".active").removeClass("active"),a=t.next();if(a.length||(a=n(this.$menu.find("li")[0])),this.options.scrollBar){var r=this.$menu.children("li").index(a);r%8==0&&this.$menu.scrollTop(26*r)}a.addClass("active")},prev:function(n){var e=this.$menu.find(".active").removeClass("active"),t=e.prev();if(t.length||(t=this.$menu.find("li").last()),this.options.scrollBar){var a=this.$menu.children("li"),r=a.length-1,i=a.index(t);(r-i)%8==0&&this.$menu.scrollTop(26*(i-7))}t.addClass("active")},listen:function(){this.$element.on("focus",n.proxy(this.focus,this)).on("blur",n.proxy(this.blur,this)).on("keypress",n.proxy(this.keypress,this)).on("keyup",n.proxy(this.keyup,this)),this.eventSupported("keydown")&&this.$element.on("keydown",n.proxy(this.keydown,this)),this.$menu.on("click",n.proxy(this.click,this)).on("mouseenter","li",n.proxy(this.mouseenter,this)).on("mouseleave","li",n.proxy(this.mouseleave,this))},move:function(n){if(this.shown){switch(n.keyCode){case 9:case 13:case 27:n.preventDefault();break;case 38:n.preventDefault(),this.prev();break;case 40:n.preventDefault(),this.next()}n.stopPropagation()}},keydown:function(e){this.suppressKeyPressRepeat=~n.inArray(e.keyCode,[40,38,9,13,27]),this.move(e)},keypress:function(n){this.suppressKeyPressRepeat||this.move(n)},keyup:function(n){switch(n.keyCode){case 40:case 38:case 16:case 17:case 18:break;case 9:case 13:if(!this.shown)return;this.select();break;case 27:if(!this.shown)return;this.hide();break;default:this.ajax?this.ajaxLookup():this.lookup()}n.stopPropagation(),n.preventDefault()},focus:function(n){this.focused=!0},blur:function(n){this.focused=!1,!this.mousedover&&this.shown&&this.hide()},click:function(n){n.stopPropagation(),n.preventDefault(),this.select(),this.$element.focus()},mouseenter:function(e){this.mousedover=!0,this.$menu.find(".active").removeClass("active"),n(e.currentTarget).addClass("active")},mouseleave:function(n){this.mousedover=!1,!this.focused&&this.shown&&this.hide()},destroy:function(){this.$element.off("focus",n.proxy(this.focus,this)).off("blur",n.proxy(this.blur,this)).off("keypress",n.proxy(this.keypress,this)).off("keyup",n.proxy(this.keyup,this)),this.eventSupported("keydown")&&this.$element.off("keydown",n.proxy(this.keydown,this)),this.$menu.off("click",n.proxy(this.click,this)).off("mouseenter","li",n.proxy(this.mouseenter,this)).off("mouseleave","li",n.proxy(this.mouseleave,this)),this.$element.removeData("typeahead")}},n.fn.typeahead=function(a){return this.each(function(){var r=n(this),i=r.data("typeahead"),o="object"==(void 0===a?"undefined":t(a))&&a;i||r.data("typeahead",i=new e(this,o)),"string"==typeof a&&i[a]()})},n.fn.typeahead.defaults={source:[],items:10,scrollBar:!1,alignWidth:!0,menu:'<ul class="typeahead dropdown-menu"></ul>',item:'<li><a href="#"></a></li>',valueField:"id",displayField:"name",autoSelect:!0,onSelect:function(){},ajax:{url:null,timeout:300,method:"get",triggerLength:1,loadingClass:null,preDispatch:null,preProcess:null}},n.fn.typeahead.Constructor=e,n(function(){n("body").on("focus.typeahead.data-api",'[data-provide="typeahead"]',function(e){var t=n(this);t.data("typeahead")||(e.preventDefault(),t.typeahead(t.data()))})})}(window.jQuery)},hcgY:function(n,e){$(function(){($("body").hasClass("program-new")||$("body").hasClass("program-edit"))&&(grantmetrics.application.setupAddRemove("program","organizer"),grantmetrics.application.setupAutocompletion()),$(".program-action__delete, .event-action__delete").on("click",function(n){return $(this).hasClass("disabled")?(n.preventDefault(),window.alert($.i18n("error-"+$(this).data("model")+"-undeletable"))):window.confirm($.i18n("confirm-deletion",$(this).data("title")))}),grantmetrics.application.setupAutocompletion(),grantmetrics.application.setupColumnSorting()})},iAVh:function(n,e){grantmetrics={},grantmetrics.application={},function(){var n={};n[i18nLang]=i18nPath,"en"!==i18nLang&&(n.en=i18nEnPath),$.i18n({locale:i18nLang}).load(n)}(),$(function(){$(".panel .panel-title a").attr("title",$.i18n("hide-section")),$(".panel").on("hidden.bs.collapse",function(n){$(this).find(".panel-title a").attr("title",$.i18n("show-section"))}).on("shown.bs.collapse",function(n){$(this).find(".panel-title a").attr("title",$.i18n("hide-section"))})}),grantmetrics.application.setupAddRemove=function(n,e){var t=$("."+n+"__"+e+"s ."+e+"-row").length,a="."+e+"-row";$(".remove-"+e).on("click",function(n){n.preventDefault(),$(this).parents(a).remove()}),$(".add-"+e).on("click",function(n){n.preventDefault();var r=$("<div class='form-group "+e+"-row'>"+$(a+"__template").html()+"</div>");$(a+":last").after(r);var i=$(a+":last");i.find("input").prop("name","form["+e+"s]["+t+"]").prop("id","form_"+e+"s_"+t).val(""),i.find(".invalid-input").remove(),t++,i.find(".remove-"+e).on("click",function(){i.remove()}),i.find("input").hasClass("user-input")&&grantmetrics.application.setupAutocompletion($(a+":last").find("input"))})},grantmetrics.application.setupAutocompletion=function(n){if(void 0===n)var n=$(".user-input");n[0]&&n.one("focus",function(){$(this).data("typeahead")&&$(this).data("typeahead").destroy();var e={url:"https://meta.wikimedia.org/w/api.php",timeout:200,triggerLength:1,method:"get"};n.typeahead({ajax:Object.assign(e,{preDispatch:function(n){return n=n.charAt(0).toUpperCase()+n.slice(1),{action:"query",list:"allusers",format:"json",aufrom:n,origin:"*"}},preProcess:function(n){return n.query.allusers.map(function(n){return n.name})}})}),$(this).trigger("focus")})},grantmetrics.application.setupColumnSorting=function(){var n,e;$(".sort-link").on("click",function(){n=e===$(this).data("column")?-n:1,$(".sort-link").removeClass("sort-link--asc sort-link--desc"),$(this).addClass(1===n?"sort-link--asc":"sort-link--desc"),e=$(this).data("column");var t=$(this).parents("table"),a=t.find(".sort-entry--"+e).parent();a.length&&(a.sort(function(t,a){var r=$(t).find(".sort-entry--"+e).data("value"),i=$(a).find(".sort-entry--"+e).data("value");return isNaN(parseFloat(r,10))||(r=parseFloat(r,10),i=parseFloat(i,10)),r<i?n:r>i?-n:0}),t.find("tbody").html($(a)))})},$(function(){function n(){var n=$(this).val().replace(/(?:[\uD800-\uDBFF][\uDC00-\uDFFF])/g,String.fromCodePoint(65533));$(this).val(n)}$("input, textarea").on({keydown:n,change:n})})},oAVT:function(n,e){$(function(){$("body").hasClass("eventdata")&&$(".event-metric-desc").tooltip()})},tSWT:function(n,e){grantmetrics.events={},$(function(){if($("body").hasClass("event")){($("body").hasClass("event-new")||$("body").hasClass("event-edit")||$("body").hasClass("event-copy"))&&grantmetrics.application.setupAddRemove("event","wiki"),$("body").hasClass("event-show")&&grantmetrics.application.setupAddRemove("event","participant");var n=moment($("#form_start").val()).utc(),e=moment($("#form_end").val()).utc();if(n=n.isValid()?n:moment().add(7,"days").startOf("week"),e=e.isValid()?e:moment().add(7,"days").endOf("week"),$("#form_time").daterangepicker({timePicker:!0,timePicker24Hour:grantmetrics.dateLocales.is24HourFormat(),startDate:n,endDate:e,locale:{format:grantmetrics.dateLocales.getLocaleDatePattern()+" "+grantmetrics.dateLocales.getLocaleTimePattern(),applyLabel:$.i18n("apply"),cancelLabel:$.i18n("cancel"),customRangeLabel:$.i18n("custom-range"),daysOfWeek:grantmetrics.dateLocales.getWeekdayNames(),monthNames:grantmetrics.dateLocales.getMonthNames()}}),$("body").hasClass("event-new")){try{timezone=Intl.DateTimeFormat().resolvedOptions().timeZone,$("#form_timezone option[value='"+timezone+"']").length&&$("#form_timezone").val(timezone)}catch(n){}}$("#event_form").on("submit",function(){var n=$("#form_time").data().daterangepicker;$("#form_start").val(n.startDate.format("YYYY-MM-DDTHH:mm:00-00:00")),$("#form_end").val(n.endDate.format("YYYY-MM-DDTHH:mm:00-00:00"))}),grantmetrics.events.populateValidWikis().then(function(n){$(".event__wikis").on("focus",".event-wiki-input",function(){$(this).data().typeahead||$(this).typeahead({source:n})})}),$(".special-wiki").on("click",function(n){n.preventDefault();var e=$(".event__wikis").find("input").last();""!==e.val().trim()&&($(".add-wiki").trigger("click"),e=$(".event__wikis").find("input").last()),e.val($(n.target).data("value"))}),$(".event-process-btn").on("click",function(){document.activeElement.blur(),$(this).addClass("disabled").text($.i18n("updating")),$(".event-export-btn").addClass("disabled"),$(".event-stats-status").text($.i18n("updating-desc")),$.get(baseUrl+"events/process/"+$(this).data("event-id")).done(function(n){$(this).removeClass("disabled").text($.i18n("update-data")),$(".event-export-btn").removeClass("disabled"),$(".event-stats-status").text(""),window.location.reload(!0)}.bind(this)).fail(function(n){$(".event-process-btn").text($.i18n("error-failed")).addClass("btn-danger");$(".event-stats-status").html("<strong class='text-danger'>"+$.i18n("error-internal","<a target='_blank' href='https://meta.wikimedia.org/wiki/Talk:Grant_Metrics'>meta:Talk:Grant Metrics</a>")+"</strong>")})}),$(".event-process-link").on("click",function(n){$(".event-process-btn").trigger("click"),$(".event-wiki-stats--empty").html("&nbsp;"),n.preventDefault()}),grantmetrics.application.setupAutocompletion(),grantmetrics.application.setupColumnSorting(),$('[data-toggle="tooltip"]').tooltip()}}),grantmetrics.events.populateValidWikis=function(){var n=$.Deferred();return $.ajax({url:"https://meta.wikimedia.org/w/api.php?action=sitematrix&formatversion=2&smsiteprop=url&smlangprop=site&format=json",dataType:"jsonp"}).then(function(e){delete e.sitematrix.count;var t=[];for(var a in e.sitematrix){var r=e.sitematrix[a];r.site&&r.site.forEach(function(n){n.closed||-1===n.url.indexOf(".wikipedia.org")||t.push(n.url.replace(/\.org$/,"").replace(/^https?:\/\//,""))})}t.push("*.wikipedia"),t.push("commons.wikimedia","www.wikidata"),n.resolve(t)}),n}},xbfo:function(n,e){$(function(){$("body").hasClass("default")&&$.get(baseUrl+"api/background/"+$(window).outerWidth()).done(function(n){$("body.default-index").attr("style","background-image:url("+n.thumburl+");"),$("#background_commons_link").text(n.canonicaltitle).prop("href",n.descriptionurl)})})}});
+!function(n){"use strict";var e=function(e,t){n.fn.typeahead.defaults,t.scrollBar&&(t.items=100,t.menu='<ul class="typeahead dropdown-menu" style="max-height:220px;overflow:auto;"></ul>');var a=this;if(a.$element=n(e),a.options=n.extend({},n.fn.typeahead.defaults,t),a.$menu=n(a.options.menu).insertAfter(a.$element),a.eventSupported=a.options.eventSupported||a.eventSupported,a.grepper=a.options.grepper||a.grepper,a.highlighter=a.options.highlighter||a.highlighter,a.lookup=a.options.lookup||a.lookup,a.matcher=a.options.matcher||a.matcher,a.render=a.options.render||a.render,a.onSelect=a.options.onSelect||null,a.sorter=a.options.sorter||a.sorter,a.select=a.options.select||a.select,a.source=a.options.source||a.source,a.displayField=a.options.displayField||a.displayField,a.valueField=a.options.valueField||a.valueField,a.autoSelect=a.options.autoSelect||a.autoSelect,a.options.ajax){var r=a.options.ajax;"string"==typeof r?a.ajax=n.extend({},n.fn.typeahead.defaults.ajax,{url:r}):("string"==typeof r.displayField&&(a.displayField=a.options.displayField=r.displayField),"string"==typeof r.valueField&&(a.valueField=a.options.valueField=r.valueField),a.ajax=n.extend({},n.fn.typeahead.defaults.ajax,r)),a.ajax.url||(a.ajax=null),a.query=""}else a.source=a.options.source,a.ajax=null;a.shown=!1,a.listen()};e.prototype={constructor:e,eventSupported:function(n){var e=n in this.$element;return e||(this.$element.setAttribute(n,"return;"),e="function"==typeof this.$element[n]),e},select:function(){var n=this.$menu.find(".active");if(n.length){var e=n.attr("data-value"),t=this.$menu.find(".active a").text();this.$element.val(this.updater(t)).change(),this.options.onSelect&&this.options.onSelect({value:e,text:t})}return this.hide()},updater:function(n){return n},show:function(){var e=n.extend({},this.$element.position(),{height:this.$element[0].offsetHeight});if(this.$menu.css({top:e.top+e.height,left:e.left}),this.options.alignWidth){var t=n(this.$element[0]).outerWidth();this.$menu.css({width:t})}return this.$menu.show(),this.shown=!0,this},hide:function(){return this.$menu.hide(),this.shown=!1,this},ajaxLookup:function(){function e(){this.ajaxToggleLoadClass(!0),this.ajax.xhr&&this.ajax.xhr.abort();var e=this.ajax.preDispatch?this.ajax.preDispatch(t):{query:t};this.ajax.xhr=n.ajax({url:this.ajax.url,data:e,success:n.proxy(this.ajaxSource,this),type:this.ajax.method||"get",dataType:"json",headers:this.ajax.headers||{}}),this.ajax.timerId=null}var t=n.trim(this.$element.val());return t===this.query?this:(this.query=t,this.ajax.timerId&&(clearTimeout(this.ajax.timerId),this.ajax.timerId=null),!t||t.length<this.ajax.triggerLength?(this.ajax.xhr&&(this.ajax.xhr.abort(),this.ajax.xhr=null,this.ajaxToggleLoadClass(!1)),this.shown?this.hide():this):(this.ajax.timerId=setTimeout(n.proxy(e,this),this.ajax.timeout),this))},ajaxSource:function(n){this.ajaxToggleLoadClass(!1);var e,t=this;if(t.ajax.xhr)return t.ajax.preProcess&&(n=t.ajax.preProcess(n)),t.ajax.data=n,e=t.grepper(t.ajax.data)||[],e.length?(t.ajax.xhr=null,t.render(e.slice(0,t.options.items)).show()):t.shown?t.hide():t},ajaxToggleLoadClass:function(n){this.ajax.loadingClass&&this.$element.toggleClass(this.ajax.loadingClass,n)},lookup:function(n){var e,t=this;return t.ajax?void t.ajaxer():(t.query=t.$element.val(),t.query&&(e=t.grepper(t.source))?(0==e.length&&(e[0]={id:-21,name:"Result not Found"}),t.render(e.slice(0,t.options.items)).show()):t.shown?t.hide():t)},matcher:function(n){return~n.toLowerCase().indexOf(this.query.toLowerCase())},sorter:function(n){if(this.options.ajax)return n;for(var e,t=[],a=[],r=[];e=n.shift();)e.toLowerCase().indexOf(this.query.toLowerCase())?~e.indexOf(this.query)?a.push(e):r.push(e):t.push(e);return t.concat(a,r)},highlighter:function(n){var e=this.query.replace(/[\-\[\]{}()*+?.,\\\^$|#\s]/g,"\\$&");return n.replace(new RegExp("("+e+")","ig"),function(n,e){return"<strong>"+e+"</strong>"})},render:function(e){var a,r=this,i="string"==typeof r.options.displayField;return e=n(e).map(function(e,o){return"object"==(void 0===o?"undefined":t(o))?(a=i?o[r.options.displayField]:r.options.displayField(o),e=n(r.options.item).attr("data-value",o[r.options.valueField])):(a=o,e=n(r.options.item).attr("data-value",o)),e.find("a").html(r.highlighter(a)),e[0]}),r.autoSelect&&e.first().addClass("active"),this.$menu.html(e),this},grepper:function(e){var t,a,r=this,i="string"==typeof r.options.displayField;if(!(i&&e&&e.length))return null;if(e[0].hasOwnProperty(r.options.displayField))t=n.grep(e,function(n){return a=i?n[r.options.displayField]:r.options.displayField(n),r.matcher(a)});else{if("string"!=typeof e[0])return null;t=n.grep(e,function(n){return r.matcher(n)})}return this.sorter(t)},next:function(e){var t=this.$menu.find(".active").removeClass("active"),a=t.next();if(a.length||(a=n(this.$menu.find("li")[0])),this.options.scrollBar){var r=this.$menu.children("li").index(a);r%8==0&&this.$menu.scrollTop(26*r)}a.addClass("active")},prev:function(n){var e=this.$menu.find(".active").removeClass("active"),t=e.prev();if(t.length||(t=this.$menu.find("li").last()),this.options.scrollBar){var a=this.$menu.children("li"),r=a.length-1,i=a.index(t);(r-i)%8==0&&this.$menu.scrollTop(26*(i-7))}t.addClass("active")},listen:function(){this.$element.on("focus",n.proxy(this.focus,this)).on("blur",n.proxy(this.blur,this)).on("keypress",n.proxy(this.keypress,this)).on("keyup",n.proxy(this.keyup,this)),this.eventSupported("keydown")&&this.$element.on("keydown",n.proxy(this.keydown,this)),this.$menu.on("click",n.proxy(this.click,this)).on("mouseenter","li",n.proxy(this.mouseenter,this)).on("mouseleave","li",n.proxy(this.mouseleave,this))},move:function(n){if(this.shown){switch(n.keyCode){case 9:case 13:case 27:n.preventDefault();break;case 38:n.preventDefault(),this.prev();break;case 40:n.preventDefault(),this.next()}n.stopPropagation()}},keydown:function(e){this.suppressKeyPressRepeat=~n.inArray(e.keyCode,[40,38,9,13,27]),this.move(e)},keypress:function(n){this.suppressKeyPressRepeat||this.move(n)},keyup:function(n){switch(n.keyCode){case 40:case 38:case 16:case 17:case 18:break;case 9:case 13:if(!this.shown)return;this.select();break;case 27:if(!this.shown)return;this.hide();break;default:this.ajax?this.ajaxLookup():this.lookup()}n.stopPropagation(),n.preventDefault()},focus:function(n){this.focused=!0},blur:function(n){this.focused=!1,!this.mousedover&&this.shown&&this.hide()},click:function(n){n.stopPropagation(),n.preventDefault(),this.select(),this.$element.focus()},mouseenter:function(e){this.mousedover=!0,this.$menu.find(".active").removeClass("active"),n(e.currentTarget).addClass("active")},mouseleave:function(n){this.mousedover=!1,!this.focused&&this.shown&&this.hide()},destroy:function(){this.$element.off("focus",n.proxy(this.focus,this)).off("blur",n.proxy(this.blur,this)).off("keypress",n.proxy(this.keypress,this)).off("keyup",n.proxy(this.keyup,this)),this.eventSupported("keydown")&&this.$element.off("keydown",n.proxy(this.keydown,this)),this.$menu.off("click",n.proxy(this.click,this)).off("mouseenter","li",n.proxy(this.mouseenter,this)).off("mouseleave","li",n.proxy(this.mouseleave,this)),this.$element.removeData("typeahead")}},n.fn.typeahead=function(a){return this.each(function(){var r=n(this),i=r.data("typeahead"),o="object"==(void 0===a?"undefined":t(a))&&a;i||r.data("typeahead",i=new e(this,o)),"string"==typeof a&&i[a]()})},n.fn.typeahead.defaults={source:[],items:10,scrollBar:!1,alignWidth:!0,menu:'<ul class="typeahead dropdown-menu"></ul>',item:'<li><a href="#"></a></li>',valueField:"id",displayField:"name",autoSelect:!0,onSelect:function(){},ajax:{url:null,timeout:300,method:"get",triggerLength:1,loadingClass:null,preDispatch:null,preProcess:null}},n.fn.typeahead.Constructor=e,n(function(){n("body").on("focus.typeahead.data-api",'[data-provide="typeahead"]',function(e){var t=n(this);t.data("typeahead")||(e.preventDefault(),t.typeahead(t.data()))})})}(window.jQuery)},hcgY:function(n,e){$(function(){($("body").hasClass("program-new")||$("body").hasClass("program-edit"))&&(grantmetrics.application.setupAddRemove("program","organizer"),grantmetrics.application.setupAutocompletion()),$(".program-action__delete, .event-action__delete").on("click",function(n){return $(this).hasClass("disabled")?(n.preventDefault(),window.alert($.i18n("error-"+$(this).data("model")+"-undeletable"))):window.confirm($.i18n("confirm-deletion",$(this).data("title")))}),grantmetrics.application.setupAutocompletion(),grantmetrics.application.setupColumnSorting()})},iAVh:function(n,e){grantmetrics={},grantmetrics.application={},function(){var n={};n[i18nLang]=i18nPath,"en"!==i18nLang&&(n.en=i18nEnPath),$.i18n({locale:i18nLang}).load(n)}(),$(function(){$(".panel .panel-title a").attr("title",$.i18n("hide-section")),$(".panel").on("hidden.bs.collapse",function(n){$(this).find(".panel-title a").attr("title",$.i18n("show-section"))}).on("shown.bs.collapse",function(n){$(this).find(".panel-title a").attr("title",$.i18n("hide-section"))})}),grantmetrics.application.setupAddRemove=function(n,e){var t=$("."+n+"__"+e+"s ."+e+"-row").length,a="."+e+"-row";$(".remove-"+e).on("click",function(n){n.preventDefault(),$(this).parents(a).remove()}),$(".add-"+e).on("click",function(r){r.preventDefault();var i=$("<div class='form-group "+e+"-row'>"+$(a+"__template").html()+"</div>");$(a+":last").after(i);var o=$(a+":last");o.find("input").prop("name",n+"["+e+"s]["+t+"]").prop("id",n+"_"+e+"s_"+t).val(""),o.find(".invalid-input").remove(),t++,o.find(".remove-"+e).on("click",function(){o.remove()}),o.find("input").hasClass("user-input")&&grantmetrics.application.setupAutocompletion($(a+":last").find("input"))})},grantmetrics.application.setupAutocompletion=function(n){if(void 0===n)var n=$(".user-input");n[0]&&n.one("focus",function(){$(this).data("typeahead")&&$(this).data("typeahead").destroy();var e={url:"https://meta.wikimedia.org/w/api.php",timeout:200,triggerLength:1,method:"get"};n.typeahead({ajax:Object.assign(e,{preDispatch:function(n){return n=n.charAt(0).toUpperCase()+n.slice(1),{action:"query",list:"allusers",format:"json",aufrom:n,origin:"*"}},preProcess:function(n){return n.query.allusers.map(function(n){return n.name})}})}),$(this).trigger("focus")})},grantmetrics.application.setupColumnSorting=function(){var n,e;$(".sort-link").on("click",function(){n=e===$(this).data("column")?-n:1,$(".sort-link").removeClass("sort-link--asc sort-link--desc"),$(this).addClass(1===n?"sort-link--asc":"sort-link--desc"),e=$(this).data("column");var t=$(this).parents("table"),a=t.find(".sort-entry--"+e).parent();a.length&&(a.sort(function(t,a){var r=$(t).find(".sort-entry--"+e).data("value"),i=$(a).find(".sort-entry--"+e).data("value");return isNaN(parseFloat(r,10))||(r=parseFloat(r,10),i=parseFloat(i,10)),r<i?n:r>i?-n:0}),t.find("tbody").html($(a)))})},$(function(){function n(){var n=$(this).val().replace(/(?:[\uD800-\uDBFF][\uDC00-\uDFFF])/g,String.fromCodePoint(65533));$(this).val(n)}$("input, textarea").on({keydown:n,change:n})})},oAVT:function(n,e){$(function(){$("body").hasClass("eventdata")&&$(".event-metric-desc").tooltip()})},tSWT:function(n,e){grantmetrics.events={},$(function(){if($("body").hasClass("event")){($("body").hasClass("event-new")||$("body").hasClass("event-edit")||$("body").hasClass("event-copy"))&&grantmetrics.application.setupAddRemove("event","wiki"),$("body").hasClass("event-show")&&grantmetrics.application.setupAddRemove("event","participant");var n=moment($("#event_start").val()).utc(),e=moment($("#event_end").val()).utc();if(n=n.isValid()?n:moment().add(7,"days").startOf("week"),e=e.isValid()?e:moment().add(7,"days").endOf("week"),$("#event_time").daterangepicker({timePicker:!0,timePicker24Hour:grantmetrics.dateLocales.is24HourFormat(),startDate:n,endDate:e,locale:{format:grantmetrics.dateLocales.getLocaleDatePattern()+" "+grantmetrics.dateLocales.getLocaleTimePattern(),applyLabel:$.i18n("apply"),cancelLabel:$.i18n("cancel"),customRangeLabel:$.i18n("custom-range"),daysOfWeek:grantmetrics.dateLocales.getWeekdayNames(),monthNames:grantmetrics.dateLocales.getMonthNames()}}),$("body").hasClass("event-new")){try{timezone=Intl.DateTimeFormat().resolvedOptions().timeZone,$("#event_timezone option[value='"+timezone+"']").length&&$("#event_timezone").val(timezone)}catch(n){}}$("#event_form").on("submit",function(){var n=$("#event_time").data().daterangepicker;$("#event_start").val(n.startDate.format("YYYY-MM-DDTHH:mm:00-00:00")),$("#event_end").val(n.endDate.format("YYYY-MM-DDTHH:mm:00-00:00"))}),grantmetrics.events.populateValidWikis().then(function(n){$(".event__wikis").on("focus",".event-wiki-input",function(){$(this).data().typeahead||$(this).typeahead({source:n})})}),$(".special-wiki").on("click",function(n){n.preventDefault();var e=$(".event__wikis").find("input").last();""!==e.val().trim()&&($(".add-wiki").trigger("click"),e=$(".event__wikis").find("input").last()),e.val($(n.target).data("value"))}),$(".event-process-btn").on("click",function(){document.activeElement.blur(),$(this).addClass("disabled").text($.i18n("updating")),$(".event-export-btn").addClass("disabled"),$(".event-stats-status").text($.i18n("updating-desc")),$.get(baseUrl+"events/process/"+$(this).data("event-id")).done(function(n){$(this).removeClass("disabled").text($.i18n("update-data")),$(".event-export-btn").removeClass("disabled"),$(".event-stats-status").text(""),window.location.reload(!0)}.bind(this)).fail(function(n){$(".event-process-btn").text($.i18n("error-failed")).addClass("btn-danger");$(".event-stats-status").html("<strong class='text-danger'>"+$.i18n("error-internal","<a target='_blank' href='https://meta.wikimedia.org/wiki/Talk:Grant_Metrics'>meta:Talk:Grant Metrics</a>")+"</strong>")})}),$(".event-process-link").on("click",function(n){$(".event-process-btn").trigger("click"),$(".event-wiki-stats--empty").html("&nbsp;"),n.preventDefault()}),grantmetrics.application.setupAutocompletion(),grantmetrics.application.setupColumnSorting(),$('[data-toggle="tooltip"]').tooltip()}}),grantmetrics.events.populateValidWikis=function(){var n=$.Deferred();return $.ajax({url:"https://meta.wikimedia.org/w/api.php?action=sitematrix&formatversion=2&smsiteprop=url&smlangprop=site&format=json",dataType:"jsonp"}).then(function(e){delete e.sitematrix.count;var t=[];for(var a in e.sitematrix){var r=e.sitematrix[a];r.site&&r.site.forEach(function(n){n.closed||-1===n.url.indexOf(".wikipedia.org")||t.push(n.url.replace(/\.org$/,"").replace(/^https?:\/\//,""))})}t.push("*.wikipedia"),t.push("commons.wikimedia","www.wikidata"),n.resolve(t)}),n}},xbfo:function(n,e){$(function(){$("body").hasClass("default")&&$.get(baseUrl+"api/background/"+$(window).outerWidth()).done(function(n){$("body.default-index").attr("style","background-image:url("+n.thumburl+");"),$("#background_commons_link").text(n.canonicaltitle).prop("href",n.descriptionurl)})})}});
