+++ public/assets/manifest.json
@@ -1,4 +1,4 @@
 {
-  "assets/app.js": "assets/app.d33e1aee.js",
+  "assets/app.js": "assets/app.e04b02e9.js",
   "assets/app.css": "assets/app.ba344134.css"
 }
