+++ src/AppBundle/Controller/EventController.php
@@ -5,30 +5,18 @@
 
 namespace AppBundle\Controller;
 
+use AppBundle\Form\EventType;
+use AppBundle\Form\ParticipantsType;
 use AppBundle\Model\Event;
 use AppBundle\Model\EventStat;
 use AppBundle\Model\EventWiki;
 use AppBundle\Model\Participant;
 use AppBundle\Repository\EventRepository;
-use AppBundle\Repository\EventWikiRepository;
-use AppBundle\Repository\ParticipantRepository;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Component\Form\CallbackTransformer;
-use Symfony\Component\Form\Extension\Core\Type\CollectionType;
-use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\Extension\Core\Type\TimezoneType;
-use Symfony\Component\Form\Form;
 use Symfony\Component\Form\FormError;
-use Symfony\Component\Form\FormEvent;
-use Symfony\Component\Form\FormEvents;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Validator\Constraints\NotBlank;
-use Symfony\Component\Validator\Constraints\Valid;
 
 /**
  * The EventController handles showing, creating and editing events.
@@ -173,7 +161,9 @@ class EventController extends EntityController
      */
     private function handleFormSubmission(Event $event)
     {
-        $form = $this->getFormForEvent($event);
+        $form = $this->createForm(EventType::class, $event, [
+            'event' => $event,
+        ]);
         $form->handleRequest($this->request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -198,9 +188,9 @@ class EventController extends EntityController
 
     /**
      * Consolidate errors of wikis associated with the event.
-     * @param Form $form
+     * @param FormInterface $form
      */
-    private function handleEventWikiErrors(Form $form)
+    private function handleEventWikiErrors(FormInterface $form)
     {
         $numWikiErrors = count($form['wikis']->getErrors(true));
         if ($numWikiErrors > 0) {
@@ -214,163 +204,9 @@ class EventController extends EntityController
         }
     }
 
-    /**
-     * Build a form for the given event.
-     * @param Event $event
-     * @return FormInterface
-     */
-    private function getFormForEvent(Event $event)
-    {
-        $builder = $this->createFormBuilder($event)
-            ->add('title', TextType::class, [
-                'constraints' => [
-                    new NotBlank(),
-                ]
-            ])
-            ->add('wikis', CollectionType::class, [
-                'entry_type' => TextType::class,
-                'allow_add' => true,
-                'allow_delete' => true,
-                'empty_data' => '',
-                'required' => true,
-                'constraints' => [new Valid(), new NotBlank()],
-                'data' => $event->getOrphanWikisAndFamilies(),
-                'data_class' => null,
-            ])
-            ->add('time', TextType::class, [
-                'mapped' => false,
-                'attr' => [
-                    'autocomplete' => 'off',
-                ],
-            ])
-            ->add('start', DateTimeType::class, [
-                'widget' => 'single_text',
-                'required' => false,
-                'html5' => false,
-                'view_timezone' => 'UTC',
-                'constraints' => [new Valid()],
-            ])
-            ->add('end', DateTimeType::class, [
-                'widget' => 'single_text',
-                'required' => false,
-                'html5' => false,
-                'view_timezone' => 'UTC',
-                'constraints' => [new Valid()],
-            ])
-            ->add('timezone', TimezoneType::class, [
-                'choices' => $this->getTimezones(),
-                'choice_loader' => null,
-            ])
-            ->add('submit', SubmitType::class)
-            ->addEventListener(FormEvents::PRE_SUBMIT, [$this, 'onEventPreSubmit']);
-
-        $builder->get('wikis')
-            ->addModelTransformer($this->getWikiCallbackTransformer($event));
-
-        return $builder->getForm();
-    }
-
-    /**
-     * Normalize the form data before submitting.
-     * @param FormEvent $formEvent
-     */
-    public function onEventPreSubmit(FormEvent $formEvent)
-    {
-        $event = $formEvent->getData();
-
-        // Remove duplicate and blank wikis.
-        $event['wikis'] = array_filter(array_values(array_unique($formEvent->getData()['wikis'])));
-
-        $formEvent->setData($event);
-    }
-
-    /**
-     * Get options for the timezone dropdown, grouping by region and
-     * also prefixing each option with the region.
-     * @return string[]
-     */
-    private function getTimezones()
-    {
-        $timezones = [
-            'UTC' => 'UTC',
-        ];
-
-        foreach (\DateTimeZone::listIdentifiers() as $timezone) {
-            $region = str_replace('_', ' ', explode('/', $timezone)[0]);
-            $displayTimezone = str_replace('_', ' ', $timezone);
-
-            if ($region === 'UTC') {
-                continue;
-            }
-
-            if (isset($timezones[$region])) {
-                $timezones[$region][$displayTimezone] = $timezone;
-            } else {
-                $timezones[$region] = [
-                    $displayTimezone => $timezone,
-                ];
-            }
-        }
-
-        return $timezones;
-    }
-
-    /**
-     * Transform wiki data to or from the form.
-     * @param Event $event
-     * @return CallbackTransformer
-     */
-    private function getWikiCallbackTransformer(Event $event)
-    {
-        $eventWikiRepo = new EventWikiRepository($this->em);
-        $eventWikiRepo->setContainer($this->container);
-
-        return new CallbackTransformer(
-            function ($wikiObjects) {
-                // To domain names for the form, from EventWikis.
-                $wikis = $wikiObjects->toArray();
-                return array_map(function (EventWiki $wiki) {
-                    return $wiki->getDomain();
-                }, $wikis);
-            },
-            function ($wikis) use ($event, $eventWikiRepo) {
-                return array_filter(
-                    $this->normalizeEventWikis($wikis, $event, $eventWikiRepo)
-                );
-            }
-        );
-    }
-
-    /**
-     * Take the list of wikis provided by the user (enwiki, en.wikipedia, or en.wikipedia.org)
-     * and normalize them to the domain (en.wikipedia). This method then instantiates a new
-     * EventWiki if one did not already exist.
-     * @param  string[]            $wikis As retrieved by the form.
-     * @param  Event               $event
-     * @param  EventWikiRepository $eventWikiRepo
-     * @return EventWiki[]
-     */
-    private function normalizeEventWikis($wikis, Event $event, EventWikiRepository $eventWikiRepo)
-    {
-        return array_map(function ($wiki) use ($event, $eventWikiRepo) {
-            $domain = $eventWikiRepo->getDomainFromEventWikiInput($wiki);
-
-            $eventWiki = $eventWikiRepo->findOneBy([
-                'event' => $event,
-                'domain' => $domain,
-            ]);
-
-            if ($eventWiki === null) {
-                $eventWiki = new EventWiki($event, $domain);
-            }
-
-            return $eventWiki;
-        }, $wikis);
-    }
-
-    /********************
-     * PARTICIPANT FORM *
-     ********************/
+    /**************
+     * EVENT PAGE *
+     **************/
 
     /**
      * Show a specific event.
@@ -382,24 +218,25 @@ class EventController extends EntityController
      *     "programTitle" = "^(?!new|edit|delete).*$",
      *     "eventTitle" = "^(?!(new|edit|delete|revisions)$)[^\/]+"
      * })
+     * @param EventRepository $eventRepo
      * @return Response
      */
     public function showAction(EventRepository $eventRepo)
     {
-        // Handle the Form for the request.
-        $form = $this->handleParticipantForm();
-        if ($form instanceof RedirectResponse) {
+        // Handle the participants form for the request.
+        $participantForm = $this->handleParticipantForm();
+        if ($participantForm instanceof RedirectResponse) {
             // Flash message will be shown at the top of the page.
             $this->addFlash('success', [
                 'event-updated',
                 $this->event->getDisplayTitle(),
             ]);
-            return $form;
+            return $participantForm;
         }
 
         return $this->render('events/show.html.twig', [
             'gmTitle' => $this->event->getDisplayTitle(),
-            'form' => $form->createView(),
+            'participantForm' => $participantForm->createView(),
             'program' => $this->program,
             'event' => $this->event,
             'stats' => $this->getEventStats($this->event),
@@ -442,7 +279,9 @@ class EventController extends EntityController
      */
     private function handleParticipantForm()
     {
-        $form = $this->getParticipantForm($this->event);
+        $form = $this->createForm(ParticipantsType::class, $this->event, [
+            'event' => $this->event,
+        ]);
         $form->handleRequest($this->request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -462,164 +301,5 @@ class EventController extends EntityController
         }
 
         return $form;
-    }
-
-    /**
-     * Get the participant list form. Shown on the 'show' page.
-     * @param  Event $event
-     * @return FormInterface
-     */
-    private function getParticipantForm(Event $event)
-    {
-        $builder = $this->createFormBuilder($event)
-            ->add('participants', CollectionType::class, [
-                'entry_type' => TextType::class,
-                'allow_add' => true,
-                'allow_delete' => true,
-                'delete_empty' => true,
-                'required' => false,
-                'constraints' => [new Valid()],
-            ])
-            ->add('new_participants', TextareaType::class, [
-                'mapped' => false,
-                'required' => false,
-            ])
-            ->add('submit', SubmitType::class)
-            ->addEventListener(FormEvents::PRE_SUBMIT, [$this, 'onParticipantPreSubmit']);
-
-        $builder->get('participants')
-            ->addModelTransformer($this->getParticipantCallbackTransformer($event));
-
-        return $builder->getForm();
-    }
-
-    /**
-     * Format data before the participant form is submitted.
-     * @param FormEvent $formEvent
-     */
-    public function onParticipantPreSubmit(FormEvent $formEvent)
-    {
-        $event = $formEvent->getData();
-
-        /**
-         * Parse new usernames from the textarea, removing
-         * carriage returns and extraneous spacing.
-         * @var string[]
-         */
-        $newParUsernames = explode("\n", $event['new_participants']);
-
-        $participants = isset($event['participants']) ? $event['participants'] : [];
-
-        // Combine usernames from inputs and textarea.
-        $event['participants'] = array_merge($participants, $newParUsernames);
-
-        // Now normalize all the usernames.
-        // TODO: Refactor this out, doing the same for Organizers to a Program.
-        // Need to somehow hook into a callback in the model layer before validations are ran.
-        $event['participants'] = array_map(function ($username) {
-            $normalized = trim(str_replace('_', ' ', str_replace("\r", '', $username)));
-
-            // Same as ucfirst but works on all locale settings. This is what MediaWiki wants.
-            return mb_strtoupper(mb_substr($normalized, 0, 1)).mb_substr($normalized, 1);
-        }, $event['participants']);
-
-        // Remove duplicates and blank entries.
-        $event['participants'] = array_filter(array_unique($event['participants']));
-
-        // Sort alphabetically.
-        sort($event['participants']);
-
-        // Now unset new_participants so they aren't duplicated in the returned form.
-        unset($event['new_participants']);
-
-        $formEvent->setData($event);
-    }
-
-    /**
-     * Given a row from ParticipantRepository::getRowsFromUsernames(),
-     * find or instantiate a new Participant.
-     * @param  Event    $event
-     * @param  string[] $row As fetched from ParticipantRepository::getRowsFromUsernames().
-     * @param  ParticipantRepository $participantRepo
-     * @return Participant
-     */
-    private function getParticipantFromRow(Event $event, $row, ParticipantRepository $participantRepo)
-    {
-        if ($row['user_id'] === null) {
-            // Username is invalid, so just return a new Participant
-            // without a user ID so that the form can produce errors.
-            $participant = new Participant($event);
-        } else {
-            // Otherwise we need find the one that exists in grantmetrics.
-            $participant = $participantRepo->findOneBy([
-                'userId' => $row['user_id'],
-                'event' => $event,
-            ]);
-
-            if ($participant === null) {
-                // Participant doesn't exist in grantmetrics yet,
-                // so we'll create a new, blank Participant.
-                $participant = new Participant($event);
-                $participant->setUserId($row['user_id']);
-            }
-        }
-
-        $participant->setUsername($row['user_name']);
-
-        return $participant;
-    }
-
-    /**
-     * Transform participant data to or from the form.
-     * This essentially pulls in the username from the user ID,
-     * and sets the user ID before persisting so that the username
-     * can be validated.
-     * @param Event $event
-     * @return CallbackTransformer
-     */
-    private function getParticipantCallbackTransformer(Event $event)
-    {
-        return new CallbackTransformer(
-            // Transform to the form.
-            function ($participantObjects) {
-                $parIds = array_map(function (Participant $participant) {
-                    return $participant->getUserId();
-                }, $participantObjects->toArray());
-
-                /** @var EventRepository $eventRepo */
-                $eventRepo = $this->em->getRepository(Event::class);
-                $eventRepo->setContainer($this->container);
-
-                $usernames = array_column($eventRepo->getUsernamesFromIds($parIds), 'user_name');
-                sort($usernames);
-                return $usernames;
-            },
-            // Transform from the form.
-            function ($participantNames) use ($event) {
-                /** @var ParticipantRepository $participantRepo */
-                $participantRepo = $this->em->getRepository(Participant::class);
-                $participantRepo->setContainer($this->container);
-
-                // Get the rows for each requested username.
-                $rows = $participantRepo->getRowsFromUsernames($participantNames);
-
-                // Should be in alphabetical order, the same order we'll show it
-                // to the user in the returned form.
-                usort($rows, function ($a, $b) {
-                    return strnatcmp($a['user_name'], $b['user_name']);
-                });
-
-                $participants = [];
-
-                // Create or get Participants from the usernames.
-                foreach ($rows as $row) {
-                    $participant = $this->getParticipantFromRow($event, $row, $participantRepo);
-                    $event->addParticipant($participant);
-                    $participants[] = $participant;
-                }
-
-                return $participants;
-            }
-        );
     }
 }
